+++ client/src/pages/Account/Account.tsx
@@ -14,6 +14,8 @@ import { CanceledError } from 'axios';
 import useAuth from '../../hooks/useAuth';
 import AccountOngoingRequestsProvider from './providers/AccountOngoingRequestsProvider';
 import AccountSocialDetailsProvider from './providers/AccountSocialDetailsProvider';
+import AccountNotificationsProvider from './providers/AccountNotificationsProvider';
+import { clearAccountNotificationsSubscriptions } from '../../services/websockets/accountNotificationsWebsSocket';
 
 export default function Account(): JSX.Element {
   const [isLoaded, setIsLoaded] = useState<boolean>(false);
@@ -64,6 +66,10 @@ export default function Account(): JSX.Element {
     return () => abortController.abort();
   }, [setAuthStatus, handleAsyncError]);
 
+  useEffect(() => {
+    return clearAccountNotificationsSubscriptions;
+  });
+
   return (
     <>
       <Head title='Account - Muninnfy' />
@@ -81,7 +87,9 @@ export default function Account(): JSX.Element {
                   initialOngoingAccountDeletionRequest={initialOngoingAccountDeletionRequest}
                 >
                   <AccountSocialDetailsProvider>
-                    <AccountContent />
+                    <AccountNotificationsProvider>
+                      <AccountContent />
+                    </AccountNotificationsProvider>
                   </AccountSocialDetailsProvider>
                 </AccountOngoingRequestsProvider>
               </AccountDetailsProvider>