+++ client/src/pages/Account/AccountSocial/AccountSocial.tsx
@@ -1,4 +1,4 @@
-import { JSX, useEffect } from 'react';
+import { JSX, useCallback, useEffect } from 'react';
 import AccountSocialHeader from './components/AccountSocialHeader/AccountSocialHeader';
 import useHandleAsyncError, { HandleAsyncErrorFunction } from '../../../hooks/useHandleAsyncError';
 import useAccountSocialDetails from '../hooks/useAccountSocialDetails';
@@ -12,6 +12,9 @@ import AccountSocialFollowing from './components/AccountSocialFollowing/AccountS
 import AccountSocialFollowRequests from './components/AccountSocialFollowRequests/AccountSocialFollowRequests';
 import AccountSocialFindAccount from './components/AccountSocialFindAccount/AccountSocialFindAccount';
 import { SOCIAL_FETCH_BATCH_SIZE } from '../../../utils/constants/socialConstants';
+import { NotificationDetails } from '../../../types/notificationTypes';
+import { FollowDetails, FollowRequest } from '../../../types/socialTypes';
+import { subscribeToAccountNotifications } from '../../../services/websockets/accountNotificationsWebsSocket';
 
 export default function AccountSocial(): JSX.Element {
   const { socialSection } = useAccountSocial();
@@ -55,6 +58,37 @@ export default function AccountSocial(): JSX.Element {
     getSocialDetails();
     return () => abortController.abort();
   }, [fetchDetails, setFetchDetails, setSocialCounts, setFollowers, setFollowing, setFollowRequests, handleAsyncError]);
+
+  const notificationsHandler = useCallback(
+    (data: NotificationDetails) => {
+      const { notification_type, notification_data } = data;
+
+      if (notification_type === 'NEW_FOLLOW_REQUEST') {
+        setFollowRequests((prev) => [notification_data as FollowRequest, ...prev]);
+        setSocialCounts((prev) => ({ ...prev, follow_requests_count: prev.follow_requests_count + 1 }));
+
+        return;
+      }
+
+      const followDetails = notification_data as FollowDetails;
+
+      if (notification_type === 'NEW_FOLLOWER') {
+        setFollowers((prev) => [followDetails, ...prev]);
+        setSocialCounts((prev) => ({ ...prev, followers_count: prev.followers_count + 1 }));
+
+        return;
+      }
+
+      setFollowing((prev) => [followDetails, ...prev]);
+      setSocialCounts((prev) => ({ ...prev, following_count: prev.following_count + 1 }));
+    },
+    [setFollowers, setFollowing, setFollowRequests, setSocialCounts]
+  );
+
+  useEffect(() => {
+    subscribeToAccountNotifications('social', notificationsHandler);
+    // unsubscribed when Account unmounts
+  }, [notificationsHandler]);
 
   return (
     <>