+++ client/src/providers/AuthProvider.tsx
@@ -2,6 +2,7 @@ import { JSX, ReactNode, useEffect, useMemo, useState } from 'react';
 import AuthContext, { AuthContextType, AuthStatus } from '../contexts/AuthContext';
 import { checkForAuthSessionService } from '../services/authServices';
 import { CanceledError } from 'axios';
+import { connectAccountNotificationsWebSocket } from '../services/websockets/accountNotificationsWebsSocket';
 
 type AuthProviderProps = {
   children: ReactNode;
@@ -17,6 +18,10 @@ export default function AuthProvider({ children }: AuthProviderProps): JSX.Eleme
       try {
         const isValidAuthSession: boolean = (await checkForAuthSessionService(abortController.signal)).data.isValidAuthSession;
         setAuthStatus(isValidAuthSession ? 'authenticated' : 'unauthenticated');
+
+        if (isValidAuthSession) {
+          connectAccountNotificationsWebSocket();
+        }
       } catch (err: unknown) {
         if (err instanceof CanceledError) {
           return;