+++ src/routers/socialRouter.ts
@@ -11,6 +11,7 @@ import { deleteFollowRequest } from '../db/helpers/socialDbHelpers';
 import { getAuthSessionId } from '../auth/authUtils';
 import { getAccountIdByAuthSessionId } from '../db/helpers/authDbHelpers';
 import { isValidSocialFindQuery, isValidSocialQuery } from '../util/validation/socialValidation';
+import { addNotification } from '../db/helpers/notificationsDbHelpers';
 
 export const socialRouter: Router = express.Router();
 
@@ -20,18 +21,18 @@ type SocialCounts = {
   follow_requests_count: number;
 };
 
-type SocialData = {
+export type SocialData = {
   public_account_id: string;
   username: string;
   display_name: string;
 };
 
-type FollowDetails = SocialData & {
+export type FollowDetails = SocialData & {
   follow_id: number;
   follow_timestamp: number;
 };
 
-type FollowRequest = SocialData & {
+export type FollowRequest = SocialData & {
   request_id: number;
   request_timestamp: number;
 };
@@ -194,7 +195,7 @@ socialRouter.get('/followers/search', async (req: Request, res: Response) => {
       { accountId, searchQuery, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: followers as FollowDetails[] });
+    res.json(followers as FollowDetails[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -252,7 +253,7 @@ socialRouter.get('/followers/:offset', async (req: Request, res: Response) => {
       { accountId, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: followers as FollowDetails[] });
+    res.json(followers as FollowDetails[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -318,7 +319,7 @@ socialRouter.get('/following/search', async (req: Request, res: Response) => {
       { accountId, searchQuery, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: following as FollowDetails[] });
+    res.json(following as FollowDetails[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -376,7 +377,7 @@ socialRouter.get('/following/:offset', async (req: Request, res: Response) => {
       { accountId, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: following as FollowDetails[] });
+    res.json(following as FollowDetails[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -442,7 +443,7 @@ socialRouter.get('/followRequests/search', async (req: Request, res: Response) =
       { accountId, searchQuery, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: followRequests as FollowRequest[] });
+    res.json(followRequests as FollowRequest[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -500,7 +501,7 @@ socialRouter.get('/followRequests/:offset', async (req: Request, res: Response) 
       { accountId, offset, socialFetchBatchSize: SOCIAL_FETCH_BATCH_SIZE }
     );
 
-    res.json({ batch: followRequests as FollowRequest[] });
+    res.json(followRequests as FollowRequest[]);
   } catch (err: unknown) {
     console.log(err);
 
@@ -648,6 +649,7 @@ socialRouter.post('/followRequests/send', async (req: Request, res: Response) =>
       await connection.commit();
       res.json({ followId: resultSetHeader.insertId, followTimestamp: currentTimestamp });
 
+      await addNotification(followDetails.requestee_account_id, accountId, currentTimestamp, 'NEW_FOLLOWER', resultSetHeader.insertId);
       return;
     }
 
@@ -662,6 +664,8 @@ socialRouter.post('/followRequests/send', async (req: Request, res: Response) =>
 
     await connection.commit();
     res.json({ requestId: resultSetHeader.insertId, requestTimestamp: currentTimestamp });
+
+    await addNotification(followDetails.requestee_account_id, accountId, currentTimestamp, 'NEW_FOLLOW_REQUEST', resultSetHeader.insertId);
   } catch (err: unknown) {
     console.log(err);
     await connection?.rollback();
@@ -823,7 +827,7 @@ socialRouter.post('/followRequests/accept', async (req: Request, res: Response) 
       return;
     }
 
-    const followTimestamp: number = Date.now();
+    const currentTimestamp: number = Date.now();
 
     const [resultSetHeader] = await connection.execute<ResultSetHeader>(
       `INSERT INTO followers (
@@ -831,7 +835,7 @@ socialRouter.post('/followRequests/accept', async (req: Request, res: Response) 
         follower_account_Id,
         follow_timestamp
       ) VALUES (${generatePlaceHolders(3)});`,
-      [accountId, followDetails.requester_account_id, followTimestamp]
+      [accountId, followDetails.requester_account_id, currentTimestamp]
     );
 
     const followRequestDeleted: boolean = await deleteFollowRequest(requestId, connection, req);
@@ -843,7 +847,15 @@ socialRouter.post('/followRequests/accept', async (req: Request, res: Response) 
     }
 
     await connection.commit();
-    res.json({ follow_id: resultSetHeader.insertId, follow_timestamp: followTimestamp });
+    res.json({ follow_id: resultSetHeader.insertId, follow_timestamp: currentTimestamp });
+
+    await addNotification(
+      followDetails.requester_account_id,
+      accountId,
+      currentTimestamp,
+      'FOLLOW_REQUEST_ACCEPTED',
+      resultSetHeader.insertId
+    );
   } catch (err: unknown) {
     console.log(err);
     await connection?.rollback();