+++ src/server.ts
@@ -1,10 +1,14 @@
-import http from 'http';
+import http, { IncomingMessage } from 'http';
 import { app } from './app';
 import { initDb } from './db/initDb';
 import { initCronJobs } from './cron-jobs/cronInit';
+import { Socket } from 'net';
+import { handleWebSocketUpgrade } from './webSocket/webSocketAuth';
 
 const port: number = process.env.PORT ? +process.env.PORT : 5000;
 const server = http.createServer(app);
+
+server.on('upgrade', async (req: IncomingMessage, socket: Socket, head: Buffer) => await handleWebSocketUpgrade(req, socket, head));
 
 async function initServer(): Promise<void> {
   try {