+++ src/util/constants/notificationsConstants.ts
@@ -1,0 +1,1 @@
+export const NOTIFICATIONS_FETCH_BATCH_SIZE: number = 100;