+++ src/util/constants/webSocketConstants.ts
@@ -1,0 +1,3 @@
+import { hourMilliseconds } from './globalConstants';
+
+export const WEB_SOCKET_INACTIVITY_THRESHOLD: number = hourMilliseconds;